Support text-decoration.
